Move HexCoordinates conversion methods as extensions
