+++ src/commonMain/kotlin/io/github/offlinebrain/khexagon/coordinates/Grid.kt
@@ -1,0 +1,5 @@
+package io.github.offlinebrain.khexagon.coordinates
+
+class Grid {
+
+}
